requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
groq==0.13.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.163.0
python-dotenv==1.0.1
python-multipart==0.0.20
itsdangerous==2.2.0
httpx==0.28.1
